Rewrite a large portion of toxcore/tox.* to be more independent
First x% is done. The primary intent of this rewrite is to pull as much
of the network code OUT of Messenger.*

There's a lot of good networking code, but all of it depends on
the Messenger. Other functions depend on the same code, but nothing
else from toxcore/Messenger.c. E.g. new groupchats could in theory
function completely independently from the messenger. From this branch
Multi-Device could function without Messenger at all. But it would
need the Networking, net_crypto, crypto_core, tox_conn, DHT, and TCP.

The two options were to piggpback of the setups that Messenger create
or to create my own. I don't need a new/separate DHT implementation,
nor a new onion. But I do need my own tox_conn. So I chose to separate
as much as I can, and use the existing tox_*() wrappers to do the work
of Messenger.c once, and allow submodules to decide what they want to
do themselves, or what they can share with toxcore proper.